style: 컴포넌트 분리 및 파일 정리
